+++ CHANGELOG.md
@@ -73,3 +73,11 @@
 ## [1.2.5]
 ### Fixed
 - Fix ui.
+
+## [1.2.6]
+### Fixed
+- GitLab project list.
+
+## [1.2.7]
+### Fixed
+- Supports batch search 4 project name search. e.g. [project-aaa,project-bbb]